chainpulse-backend/requirements.txt
fastapi
uvicorn
python-socketio
redis
psycopg2-binary
sqlalchemy
asyncpg
pandas
numpy
scikit-learn
xgboost
joblib
textblob
httpx
python-dotenv
pydantic
